Configuração de conexão com banco de dados com carga padrão CSV na inicialização [via Seed Migrations]

+++ CityGovernance.Domain/Models/City.cs
@@ -14,13 +14,14 @@ namespace CityGovernance.Domain.Models
 
         public string Name { get;  set; }
 
-        public float Longitude { get;  set; }
+        public double Longitude { get;  set; }
 
-        public float Latitude { get;  set; }
+        public double Latitude { get;  set; }
 
         public string Uf { get;  set; }
 
         public Region Region { get; set; }
+
 
     }
 }
